% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boston_athlete.R
\name{boston_athlete}
\alias{boston_athlete}
\title{Get basic data of an athlete that finish Boston Marathon.}
\usage{
boston_athlete(athlete)
}
\arguments{
\item{athlete}{Year when the event occured (string).}
}
\value{
A ggplot2 plot with density of the athletes (y-axis) over
finishing times (x-axis).
}
\description{
The boston_athlete() function uses the results_boston database to get information
about an athlete that finish Boston Marathon from 2009 to 2022.
}
\examples{
boston_athlete(athlete = "Kipruto")
boston_athlete(athlete = "Benson Kipruto")

}
