% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boston_country.R
\name{boston_country}
\alias{boston_country}
\title{Get summary information about countries in Boston Marathon.}
\usage{
boston_country(gender, year, country)
}
\arguments{
\item{gender}{Gender of the athlete (string).}

\item{year}{Year when the event occured (string).}

\item{country}{Country of origin of an athlete (string).}
}
\value{
A dataframe of 7 column with gender, year of event, country of origin
of an athlete, number of finishers from that country, mean age of finishers,
fastest time and slowest time.
}
\description{
The boston_country() function uses the results_boston database to get information
about countries that had representants in an edition of Boston Marathon from 2009 to 2022.
}
\examples{
boston_country(gender = "both", year = 2021:2022, country = c("Brazil", "Kenya"))
boston_country(gender = "Female", year = 2022, country = "Kenya")
boston_country(gender = "Male", year = 2009, country = "United States of America")

}
