% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boston_density_ridges.R
\name{boston_density_ridges}
\alias{boston_density_ridges}
\title{Plot density of participants over finishing time in Boston Marathon.}
\usage{
boston_density_ridges(year, range_position)
}
\arguments{
\item{year}{Year when the event occured (string).}

\item{range_position}{Range of positions to be ploted (string).}
}
\value{
A dataframe of 5 column with year of event, full name of the athlete,
age of race day, finishing position by gender and official.
}
\description{
The boston_density_ridges() function uses the results_boston database to plot
a density plot (based on the package "ggridges") of finishing times of
Boston Marathon from 2009 to 2022. This function needs at least two years to
for the graph to plotted properly.
}
\examples{
boston_density_ridges(year = 2012:2019, range_position = 1:10)

}
