% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boston_position.R
\name{boston_position}
\alias{boston_position}
\title{Find who finished in a certain position of Boston Marathon}
\usage{
boston_position(gender, year, position)
}
\arguments{
\item{gender}{Gender of the athlete (string).}

\item{year}{Year when the event occured (string).}

\item{position}{Position which the athlete finished the race (integer).}
}
\value{
A dataframe of 7 column with year of event, name of the athlete,
age of race day, gender, country of citizenship, official time and
finishing rank by gender.
}
\description{
The boston_position() function uses the results_boston database to find who
finished in a certain position of male and female Boston Marathons from 2009 to 2022.
}
\examples{
boston_position(gender = "both", year = 2021:2022, position = 1:3)
boston_position(gender = "Female", year = c(2010, 2021:2022), position = 1)
boston_position(gender = "Male", year = 2009, position = 1:10)

}
