% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boston_summary.R
\name{boston_summary}
\alias{boston_summary}
\title{Get a summary stats of the participants of the Boston Marathon}
\usage{
boston_summary(gender, year)
}
\arguments{
\item{gender}{Gender of the athlete (string).}

\item{year}{Year when the event occured (string).}
}
\value{
A dataframe of 7 column with gender, year of event, number of finishers, mean age of finishers,
number of countries that had representants in the event, fastest time (champion)
and slowest time.
}
\description{
The boston_summary() function uses the results_boston database to get stats about
the participants of a certain edition of Boston Marathon from 2009 to 2022.
}
\examples{
boston_summary(gender = "Male", year = 2021:2022)
boston_summary(gender = "Female", year = 2021)
boston_summary(gender = "both", year = c(2009, 2021:2022))

}
