% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results_boston}
\alias{results_boston}
\title{Results of Boston Marathon}
\format{
A dataframe with:
\describe{
\item{Year}{Year of event}
\item{FullName}{Full name of the athlete}
\item{AgeOnRaceDay}{Age of the athlete on race day}
\item{Gender}{Gender of the athlete}
\item{City}{City of residence of the athlete}
\item{StateName}{State of residence of the athlete}
\item{Zip}{Zipcode of residence of the athlete}
\item{CountryOfResAbbrev}{Abbreviation of country of residence of the athlete}
\item{CountryOfResName}{Country of residence of the athlete}
\item{CountryOfCtzAbbrev}{Abbreviation of country citizenship of the athlete}
\item{CountryOfCtzName}{Country citizenship of the athlete}
\item{OfficialTime}{Official finihsing time of the athlete}
\item{RankOverall}{Overal finishing rank of the athlete}
\item{RankOverGender}{Finishing rank over gender of the athlete}
\item{RankOverDivision}{Finishing rank over division of the athlete}
\item{SubGroupLabel}{Subgroup label of the athlete}
\item{SubGroup}{Subgroup description of the athlete}
}
}
\source{
\url{http://registration.baa.org}
}
\usage{
results_boston
}
\description{
Results of finishers of Boston Marathon from 2009 to 2022.
}
\keyword{datasets}
